Added less featurization system.
process and mprocess can transform filenames or arrays into features.
They can process using a list of featurizers which satisfy a simple
interface and output pandas.

Note that the old distance featurization outputted pandas that were
solely composed of features, while these functions output tables with
identifiers for each row and features. This will have to be adjusted for
in scripts, although the old distance functions still are backwards
compatible.