Fixed ordering bug in distance calculation, added sorting interface.
Commit 2f6a9106eb656b7ea5e63cb1c05c2828eddcd1b7 introduced a bug and did
not solve the targeted bug. Distances returned via distance function
were incorrect. They are now correct and the intended behavior of
sorting the individual labels in each column name can be performed.

Additionally, this interface is used in the command line tools to have
sorted distance labels in the produced csvs.